third_party/WebKit/LayoutTests/fast/css/radial-gradient-calc-percentage-pixels.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_false(CSS.supports('background', 'radial-gradient(calc(60% + 5px) circle at right calc(60px + 5px) top 50%, yellow, blue)'), 'rejects calcs with percentage and pixels');
  assert_false(CSS.supports('background', 'radial-gradient(calc(60% + 5px) ellipse at right calc(60px + 5px) top 50%, yellow, blue)'), 'rejects calcs with percentage and pixels');
  assert_false(CSS.supports('background', 'radial-gradient(calc(60% + 5px) calc(60% + 5px) ellipse at right calc(60px + 5px) top 50%, yellow, blue)'), 'rejects calcs with percentage and pixels');
  assert_true(CSS.supports('background', 'radial-gradient(60% 5px ellipse at right calc(60px + 5px) top 50%, yellow, blue)'), 'rejects calcs with percentage and pixels');
}, 'radial-gradient with calcs using percentage and pixels.');
</script>
